feat(basic-packages): install and configure redux persist
