+++ package.json
@@ -21,6 +21,7 @@
     "react-tippy": "^1.3.4",
     "redux": "^4.0.5",
     "redux-i18n": "^1.5.19",
+    "redux-persist": "^6.0.0",
     "redux-thunk": "^2.3.0"
   },
   "scripts": {
