--- src/configureStore.js
@@ -1,18 +1,0 @@
-import { applyMiddleware, compose, createStore } from 'redux'
-import thunkMiddleware from 'redux-thunk'
-
-import monitorReducersEnhancer from './Enhancers/monitorReducer'
-import loggerMiddleware from './Middleware/logger'
-import rootReducer from './Reducers'
-
-export default function configureStore(preloadedState) {
-  const middlewares = [loggerMiddleware, thunkMiddleware]
-  const middlewareEnhancer = applyMiddleware(...middlewares)
-
-  const enhancers = [middlewareEnhancer, monitorReducersEnhancer]
-  const composedEnhancers = compose(...enhancers)
-
-  const store = createStore(rootReducer, preloadedState, composedEnhancers)
-
-  return store
-}
